♻️ split AgentManagementConfig.tsx file, reduce the number of rows from 1400 to 900.
