+++ frontend/app/[locale]/setup/agentSetup/AgentManagementConfig.tsx
@@ -1,639 +1,33 @@
 "use client"
 
-import { useState, useEffect, useMemo, useCallback, memo, useRef } from 'react'
-import { Typography, Input, Button, Switch, Modal, message, Select, InputNumber, Tag, Upload } from 'antd'
-import { SettingOutlined, UploadOutlined, ThunderboltOutlined, LoadingOutlined, ApiOutlined, ReloadOutlined } from '@ant-design/icons'
-import ToolConfigModal from './components/ToolConfigModal'
-import McpConfigModal from './components/McpConfigModal'
-import AgentInfoInput from './components/AgentInfoInput'
-import AgentContextMenu from './components/AgentContextMenu'
-import { Tool, SubAgentPoolProps, ToolPoolProps, BusinessLogicConfigProps, Agent, OpenAIModel } from './ConstInterface'
-import { ScrollArea } from '@/components/ui/scrollArea'
-import { getCreatingSubAgentId, fetchAgentList, updateToolConfig, searchToolConfig, updateAgent, importAgent, exportAgent, deleteAgent, searchAgentInfo, fetchTools } from '@/services/agentConfigService'
-import { updateToolList } from '@/services/mcpService'
-import {generatePromptStream, savePrompt} from '@/services/promptService'
+import { useState, useEffect, useCallback, useRef } from 'react'
+import { message, Modal } from 'antd'
 import { useTranslation } from 'react-i18next'
 import { TFunction } from 'i18next'
-import { Tooltip as CustomTooltip, TooltipTrigger, TooltipContent, TooltipProvider } from '@/components/ui/tooltip'
-const { TextArea } = Input
-
-const ModelOptions = () => {
-  const { t } = useTranslation('common');
-  return [
-    { label: t('model.option.main'), value: OpenAIModel.MainModel },
-    { label: t('model.option.sub'), value: OpenAIModel.SubModel },
-  ]
-}
-
-// 提取公共的 handleToolSelect 逻辑
-const handleToolSelectCommon = async (
-  tool: Tool,
-  isSelected: boolean,
-  mainAgentId: string | null | undefined,
-  t: TFunction,
-  onSuccess?: (tool: Tool, isSelected: boolean) => void
-) => {
-
-  // Only block the action when attempting to select an unavailable tool.
-  if (tool.is_available === false && isSelected) {
-    message.error(t('tool.message.unavailable'));
-    return { shouldProceed: false, params: {} };
-  }
-
-  if (!mainAgentId) {
-    message.error(t('tool.error.noMainAgentId'));
-    return { shouldProceed: false, params: {} };
-  }
-
-  try {
-    // step 1: get tool config from database
-    const searchResult = await searchToolConfig(parseInt(tool.id), parseInt(mainAgentId));
-    if (!searchResult.success) {
-      message.error(t('tool.error.configFetchFailed'));
-      return { shouldProceed: false, params: {} };
-    }
-
-    let params: Record<string, any> = {};
-
-    // use config from database or default config
-    if (searchResult.data?.params) {
-      params = searchResult.data.params || {};
-    } else {
-      // if there is no saved config, use default value
-      params = (tool.initParams || []).reduce((acc, param) => {
-        if (param && param.name) {
-          acc[param.name] = param.value;
-        }
-        return acc;
-      }, {} as Record<string, any>);
-    }
-
-    // step 2: if the tool is enabled, check required fields
-    if (isSelected && tool.initParams && tool.initParams.length > 0) {
-      const missingRequiredFields = tool.initParams
-        .filter(param => param && param.required && (params[param.name] === undefined || params[param.name] === '' || params[param.name] === null))
-        .map(param => param.name);
-
-      if (missingRequiredFields.length > 0) {
-        return { shouldProceed: false, params };
-      }
-    }
-
-    // step 3: if all checks pass, update tool config
-    const updateResult = await updateToolConfig(
-      parseInt(tool.id),
-      parseInt(mainAgentId),
-      params,
-      isSelected
-    );
-
-    if (updateResult.success) {
-      if (onSuccess) {
-        onSuccess(tool, isSelected);
-      }
-      message.success(t('tool.message.statusUpdated', { name: tool.name, status: isSelected ? t('common.enabled') : t('common.disabled') }));
-      return { shouldProceed: true, params };
-    } else {
-      message.error(updateResult.message || t('tool.error.updateFailed'));
-      return { shouldProceed: false, params };
-    }
-  } catch (error) {
-    message.error(t('tool.error.updateRetry'));
-    return { shouldProceed: false, params: {} };
-  }
-};
-
-/**
- * Business Logic Input Component
- */
-function BusinessLogicInput({ 
-  value, 
-  onChange
-}: {
-  value: string;
-  onChange: (value: string) => void;
-}) {
-  const { t } = useTranslation('common');
-
-  return (
-    <div className="flex flex-col h-full">
-      <div className="flex items-center mb-2">
-        <div className="flex items-center justify-center w-6 h-6 rounded-full bg-blue-500 text-white text-sm font-medium mr-2">
-          3
-        </div>
-        <h2 className="text-lg font-medium">{t('businessLogic.title')}</h2>
-      </div>
-      <div className="flex-1 flex flex-col">
-        <TextArea
-          value={value}
-          onChange={(e) => onChange(e.target.value)}
-          placeholder={t('businessLogic.placeholder')}
-          className="w-full h-full resize-none p-3 text-base"
-          style={{ height: '100%' }}
-          autoSize={false}
-        />
-      </div>
-    </div>
-  )
-}
-
-/**
- * Sub Agent Pool Component
- */
-function SubAgentPool({ 
-  selectedAgents, 
-  onSelectAgent, 
-  onEditAgent, 
-  onCreateNewAgent, 
-  onImportAgent,
-  onExportAgent,
-  onDeleteAgent,
-  subAgentList = [],
-  loadingAgents = false,
-  enabledAgentIds = [],
-  isImporting = false
-}: SubAgentPoolProps) {
-  const { t } = useTranslation('common');
-
-  const [contextMenu, setContextMenu] = useState<{
-    visible: boolean;
-    x: number;
-    y: number;
-    agent: Agent | null;
-  }>({
-    visible: false,
-    x: 0,
-    y: 0,
-    agent: null
-  });
-
-  const handleContextMenu = (e: React.MouseEvent, agent: Agent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    
-    setContextMenu({
-      visible: true,
-      x: e.clientX,
-      y: e.clientY,
-      agent: agent
-    });
-  };
-
-  const handleCloseContextMenu = () => {
-    setContextMenu({
-      visible: false,
-      x: 0,
-      y: 0,
-      agent: null
-    });
-  };
-
-  return (
-    <div className="flex flex-col h-full min-h-0 overflow-hidden">
-      <div className="flex justify-between items-center mb-2">
-        <div className="flex items-center">
-          <div className="flex items-center justify-center w-6 h-6 rounded-full bg-blue-500 text-white text-sm font-medium mr-2">
-            1
-          </div>
-          <h2 className="text-lg font-medium">{t('subAgentPool.title')}</h2>
-        </div>
-        {loadingAgents && <span className="text-sm text-gray-500">{t('subAgentPool.loading')}</span>}
-      </div>
-      <ScrollArea className="flex-1 min-h-0 border-t pt-2 pb-2">
-        <div className="grid grid-cols-1 gap-3 pr-2">
-          <div className="flex gap-2 mb-2">
-            <div 
-              className="flex-1 border rounded-md p-3 flex flex-col justify-center items-center cursor-pointer transition-colors duration-200 h-[80px] hover:border-blue-300 hover:bg-blue-50"
-              onClick={onCreateNewAgent}
-            >
-              <div className="flex items-center justify-center h-full text-blue-500">
-                <span className="text-lg mr-2">+</span>
-                <span className="text-sm">{t('subAgentPool.button.create')}</span>
-              </div>
-            </div>
-
-            <div
-              className={`flex-1 border rounded-md p-3 flex flex-col justify-center items-center transition-colors duration-200 h-[80px] ${
-                isImporting
-                  ? 'opacity-50 cursor-not-allowed border-gray-300'
-                  : 'cursor-pointer hover:border-green-300 hover:bg-green-50'
-              }`}
-              onClick={isImporting ? undefined : onImportAgent}
-            >
-              <div className={`flex items-center justify-center h-full ${isImporting ? 'text-gray-400' : 'text-green-500'}`}>
-                <UploadOutlined className="text-lg mr-2" />
-                <span className="text-sm">{isImporting ? t('subAgentPool.button.importing') : t('subAgentPool.button.import')}</span>
-              </div>
-            </div>
-          </div>
-
-          {subAgentList.map((agent) => {
-            const isEnabled = enabledAgentIds.includes(Number(agent.id));
-            const isAvailable = agent.is_available !== false; // 默认为true，只有明确为false时才不可用
-            return (
-              <div 
-                key={agent.id} 
-                className={`border rounded-md p-3 flex flex-col justify-center transition-colors duration-200 h-[80px] ${
-                  !isAvailable
-                    ? isEnabled
-                      ? 'bg-blue-100 border-blue-400 opacity-60'
-                      : 'bg-gray-50 border-gray-200 opacity-60 cursor-not-allowed'
-                    : isEnabled 
-                      ? 'bg-blue-100 border-blue-400' 
-                      : 'hover:border-blue-300'
-                }`}
-                title={!isAvailable 
-                  ? isEnabled 
-                    ? t('subAgentPool.tooltip.disabledAgent')
-                    : t('subAgentPool.tooltip.unavailableAgent')
-                  : undefined}
-                onClick={() => {
-                  if (!isAvailable && !isEnabled) {
-                    message.warning(t('subAgentPool.message.unavailable'));
-                    return;
-                  }
-                  // 如果Agent不可用但已启用，允许禁用它
-                  if (!isAvailable && isEnabled) {
-                    onSelectAgent(agent, false);
-                    return;
-                  }
-                  onSelectAgent(agent, !isEnabled);
-                }}
-                onContextMenu={(e) => handleContextMenu(e, agent)}
-              >
-                <div className="flex items-center h-full">
-                  <div className="flex-1 overflow-hidden">
-                    <div className={`font-medium text-sm truncate ${!isAvailable && !isEnabled ? 'text-gray-400' : ''}`} 
-                         title={!isAvailable 
-                           ? isEnabled 
-                             ? t('subAgentPool.tooltip.disabledAgent')
-                             : t('subAgentPool.tooltip.unavailableAgent')
-                           : agent.name}>
-                      {agent.name}
-                    </div>
-                    <div 
-                      className={`text-xs line-clamp-2 ${!isAvailable && !isEnabled ? 'text-gray-400' : 'text-gray-500'}`}
-                      title={agent.description}
-                    >
-                      {agent.description}
-                    </div>
-                  </div>
-                  <button 
-                    type="button"
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      onEditAgent(agent);
-                    }}
-                    className={`ml-2 flex-shrink-0 flex items-center justify-center bg-transparent ${
-                      !isAvailable ? 'text-gray-400 hover:text-gray-600' : 'text-gray-500 hover:text-blue-500'
-                    }`}
-                    style={{ border: "none", padding: "4px" }}
-                  >
-                    <SettingOutlined style={{ fontSize: '16px' }} />
-                  </button>
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      </ScrollArea>
-
-      {/* Context Menu */}
-      <AgentContextMenu
-        visible={contextMenu.visible}
-        x={contextMenu.x}
-        y={contextMenu.y}
-        agent={contextMenu.agent}
-        onEdit={onEditAgent}
-        onExport={onExportAgent}
-        onDelete={onDeleteAgent}
-        onClose={handleCloseContextMenu}
-      />
-    </div>
-  )
-}
-
-/**
- * Tool Pool Component
- */
-function ToolPool({ 
-  selectedTools, 
-  onSelectTool, 
-  isCreatingNewAgent, 
-  tools = [], 
-  loadingTools = false,
-  mainAgentId,
-  localIsGenerating,
-  onToolsRefresh
-}: ToolPoolProps) {
-  const { t } = useTranslation('common');
-
-  const [isToolModalOpen, setIsToolModalOpen] = useState(false);
-  const [currentTool, setCurrentTool] = useState<Tool | null>(null);
-  const [pendingToolSelection, setPendingToolSelection] = useState<{tool: Tool, isSelected: boolean} | null>(null);
-  const [isMcpModalOpen, setIsMcpModalOpen] = useState(false);
-  const [isRefreshing, setIsRefreshing] = useState(false);
-  // Use useMemo to cache the tool list to avoid unnecessary recalculations
-  const displayTools = useMemo(() => {
-    const toolsToSort = tools || [];
-    // Sort by create_time, earliest created tools first (ascending order)
-    return toolsToSort.sort((a, b) => {
-      // If either tool doesn't have create_time, treat it as newer (sort to bottom)
-      if (!a.create_time && !b.create_time) return 0;
-      if (!a.create_time) return 1;
-      if (!b.create_time) return -1;
-      
-      // Compare create_time strings (ISO format can be compared directly)
-      return a.create_time.localeCompare(b.create_time);
-    });
-  }, [tools]);
-
-  // Use useMemo to cache the selected tool ID set to improve lookup efficiency
-  const selectedToolIds = useMemo(() => {
-    return new Set(selectedTools.map(tool => tool.id));
-  }, [selectedTools]);
-
-  // Use useCallback to cache the tool selection processing function
-  const handleToolSelect = useCallback(async (tool: Tool, isSelected: boolean, e: React.MouseEvent) => {
-    e.stopPropagation();
-    
-    const { shouldProceed, params } = await handleToolSelectCommon(
-      tool,
-      isSelected,
-      mainAgentId,
-      t,
-      (tool, isSelected) => onSelectTool(tool, isSelected)
-    );
-
-    if (!shouldProceed && params) {
-      setCurrentTool({
-        ...tool,
-        initParams: tool.initParams.map(param => ({
-          ...param,
-          value: params[param.name] || param.value
-        }))
-      });
-      setPendingToolSelection({ tool, isSelected });
-      setIsToolModalOpen(true);
-    }
-  }, [mainAgentId, onSelectTool, t]);
-
-  // Use useCallback to cache the tool configuration click processing function
-  const handleConfigClick = useCallback((tool: Tool, e: React.MouseEvent) => {
-    e.stopPropagation();
-    setCurrentTool(tool);
-    setIsToolModalOpen(true);
-  }, []);
-
-  // Use useCallback to cache the tool save processing function
-  const handleToolSave = useCallback((updatedTool: Tool) => {
-    if (pendingToolSelection) {
-      const { tool, isSelected } = pendingToolSelection;
-      const missingRequiredFields = updatedTool.initParams
-        .filter(param => param.required && (param.value === undefined || param.value === '' || param.value === null))
-        .map(param => param.name);
-
-      if (missingRequiredFields.length > 0) {
-        message.error(t('toolPool.error.requiredFields', { fields: missingRequiredFields.join(', ') }));
-        return;
-      }
-
-      const mockEvent = {
-        stopPropagation: () => {},
-        preventDefault: () => {},
-        nativeEvent: new MouseEvent('click'),
-        isDefaultPrevented: () => false,
-        isPropagationStopped: () => false,
-        persist: () => {}
-      } as React.MouseEvent;
-      
-      handleToolSelect(updatedTool, isSelected, mockEvent);
-    }
-    
-    setIsToolModalOpen(false);
-    setPendingToolSelection(null);
-  }, [pendingToolSelection, handleToolSelect, t]);
-
-  // Use useCallback to cache the modal close processing function
-  const handleModalClose = useCallback(() => {
-    setIsToolModalOpen(false);
-    setPendingToolSelection(null);
-  }, []);
-
-  // 刷新工具列表的处理函数
-  const handleRefreshTools = useCallback(async () => {
-    if (isRefreshing || localIsGenerating) return;
-
-    setIsRefreshing(true);
-    try {
-      // 第一步：更新后端工具状态，重新扫描MCP和本地工具
-      const updateResult = await updateToolList();
-      if (!updateResult.success) {
-        message.warning(t('toolManagement.message.updateStatusFailed'));
-      }
-
-      // 第二步：获取最新的工具列表
-      const fetchResult = await fetchTools();
-      if (fetchResult.success) {
-        // 调用父组件的刷新回调，更新工具列表状态
-        if (onToolsRefresh) {
-          onToolsRefresh();
-        }
-      } else {
-        message.error(fetchResult.message || t('toolManagement.message.refreshFailed'));
-      }
-    } catch (error) {
-      console.error(t('debug.console.refreshToolsFailed'), error);
-      message.error(t('toolManagement.message.refreshFailedRetry'));
-    } finally {
-      setIsRefreshing(false);
-    }
-  }, [isRefreshing, localIsGenerating, onToolsRefresh, t]);
-
-
-
-  // 监听工具更新事件
-  useEffect(() => {
-    const handleToolsUpdate = async () => {
-      try {
-        // 重新获取最新的工具列表，确保包含新添加的MCP工具
-        const fetchResult = await fetchTools();
-        if (fetchResult.success) {
-          // 调用父组件的刷新回调，更新工具列表状态
-          if (onToolsRefresh) {
-            onToolsRefresh();
-          }
-        } else {
-          console.error('MCP配置后自动刷新工具列表失败:', fetchResult.message);
-        }
-      } catch (error) {
-        console.error('MCP配置后自动刷新工具列表出错:', error);
-      }
-    };
-
-    window.addEventListener('toolsUpdated', handleToolsUpdate);
-    return () => {
-      window.removeEventListener('toolsUpdated', handleToolsUpdate);
-    };
-  }, [onToolsRefresh]);
-
-  // Use memo to optimize the rendering of tool items
-  const ToolItem = memo(({ tool }: { tool: Tool }) => {
-    const isSelected = selectedToolIds.has(tool.id);
-    const isAvailable = tool.is_available !== false; // 默认为true，只有明确为false时才不可用
-    
-    return (
-      <div 
-        className={`border rounded-md p-2 flex items-center transition-colors duration-200 min-h-[45px] ${
-          !isAvailable
-            ? isSelected
-              ? 'bg-blue-100 border-blue-400 opacity-60'
-              : 'bg-gray-50 border-gray-200 opacity-60 cursor-not-allowed'
-            : isSelected 
-              ? 'bg-blue-100 border-blue-400' 
-              : 'hover:border-blue-300'
-        } ${localIsGenerating ? 'opacity-50 cursor-not-allowed' : 'cursor-pointer'}`}
-        title={!isAvailable 
-          ? isSelected 
-            ? t('toolPool.tooltip.disabledTool')
-            : t('toolPool.tooltip.unavailableTool')
-          : tool.name}
-        onClick={(e) => {
-          if (localIsGenerating) return;
-          if (!isAvailable && !isSelected) {
-            message.warning(t('toolPool.message.unavailable'));
-            return;
-          }
-          handleToolSelect(tool, !isSelected, e);
-        }}
-      >
-        {/* Tool name left */}
-        <div className="flex-1 overflow-hidden">
-          <CustomTooltip>
-            <TooltipTrigger asChild>
-              <div
-                className={`font-medium text-sm truncate ${!isAvailable && !isSelected ? 'text-gray-400' : ''}`}
-                style={{
-                  maxWidth: '300px',
-                  overflow: 'hidden',
-                  textOverflow: 'ellipsis',
-                  whiteSpace: 'nowrap',
-                  display: 'inline-block',
-                  verticalAlign: 'middle'
-                }}
-              >
-                {tool.name}
-              </div>
-            </TooltipTrigger>
-            <TooltipContent side="top">
-              {tool.name}
-            </TooltipContent>
-          </CustomTooltip>
-        </div>
-        {/* Tag and settings button right */}
-        <div className="flex items-center gap-2 ml-2">
-          <div className="flex items-center justify-start min-w-[90px] w-[90px]">
-            <Tag color={tool?.source === 'mcp' ? 'blue' : 'green'} 
-                 className={`w-full text-center ${!isAvailable && !isSelected ? 'opacity-50' : ''}`}>
-              {tool?.source === 'mcp' ? t('toolPool.tag.mcp') : t('toolPool.tag.local')}
-            </Tag>
-          </div>
-          <button 
-            type="button"
-            onClick={(e) => {
-              e.stopPropagation();  // 防止触发父元素的点击事件
-              if (localIsGenerating) return;
-              if (!isAvailable) {
-                if (isSelected) {
-                  handleToolSelect(tool, false, e);
-                } else {
-                  message.warning(t('toolPool.message.unavailable'));
-                }
-                return;
-              }
-              handleConfigClick(tool, e);
-            }}
-            disabled={localIsGenerating}
-            className={`flex-shrink-0 flex items-center justify-center bg-transparent ${
-              localIsGenerating ? 'text-gray-300 cursor-not-allowed' : 'text-gray-500 hover:text-blue-500'
-            }`}
-            style={{ border: "none", padding: "4px" }}
-          >
-            <SettingOutlined style={{ fontSize: '16px' }} />
-          </button>
-        </div>
-      </div>
-    );
-  });
-
-  return (
-    <div className="flex flex-col h-full min-h-0 overflow-hidden">
-      <div className="flex justify-between items-center mb-2">
-        <div className="flex items-center">
-          <div className="flex items-center justify-center w-6 h-6 rounded-full bg-blue-500 text-white text-sm font-medium mr-2">
-            2
-          </div>
-          <h2 className="text-lg font-medium">{t('toolPool.title')}</h2>
-        </div>
-        <div className="flex items-center gap-2">
-          <Button
-            type="text"
-            size="small"
-            icon={isRefreshing ? <LoadingOutlined /> : <ReloadOutlined />}
-            onClick={handleRefreshTools}
-            disabled={localIsGenerating || isRefreshing}
-            className="text-green-500 hover:text-green-600 hover:bg-green-50"
-            title={t('toolManagement.refresh.title')}
-          >
-            {isRefreshing ? t('toolManagement.refresh.button.refreshing') : t('toolManagement.refresh.button.refresh')}
-          </Button>
-          <Button
-            type="text"
-            size="small"
-            icon={<ApiOutlined />}
-            onClick={() => setIsMcpModalOpen(true)}
-            disabled={localIsGenerating}
-            className="text-blue-500 hover:text-blue-600 hover:bg-blue-50"
-            title={t('toolManagement.mcp.title')}
-          >
-            {t('toolManagement.mcp.button')}
-          </Button>
-          {loadingTools && <span className="text-sm text-gray-500">{t('toolPool.loading')}</span>}
-        </div>
-      </div>
-      <ScrollArea className="flex-1 min-h-0 border-t pt-2 pb-2">
-        {loadingTools ? (
-          <div className="flex items-center justify-center h-full">
-            <span className="text-gray-500">{t('toolPool.loadingTools')}</span>
-          </div>
-        ) : (
-          <div className="flex flex-col gap-3 pr-2">
-            {displayTools.map((tool) => (
-              <ToolItem key={tool.id} tool={tool} />
-            ))}
-          </div>
-        )}
-      </ScrollArea>
-
-      <ToolConfigModal
-        isOpen={isToolModalOpen}
-        onCancel={handleModalClose}
-        onSave={handleToolSave}
-        tool={currentTool}
-        mainAgentId={parseInt(mainAgentId || '0')}
-        selectedTools={selectedTools}
-      />
-
-      <McpConfigModal
-        visible={isMcpModalOpen}
-        onCancel={() => setIsMcpModalOpen(false)}
-      />
-    </div>
-  );
-}
-
-// Use memo to optimize the rendering of ToolPool component
-export const MemoizedToolPool = memo(ToolPool);
+import { TooltipProvider } from '@/components/ui/tooltip'
+import BusinessLogicInput from './components/BusinessLogicInput'
+import SubAgentPool from './components/SubAgentPool'
+import { MemoizedToolPool } from './components/ToolPool'
+import ModelConfigPanel from './components/ModelConfigPanel'
+import ActionButtons from './components/ActionButtons'
+import DeleteConfirmModal from './components/DeleteConfirmModal'
+import AgentInfoInput from './components/AgentInfoInput'
+import { 
+  BusinessLogicConfigProps, 
+  Agent, 
+  Tool, 
+  OpenAIModel 
+} from './ConstInterface'
+import { 
+  getCreatingSubAgentId, 
+  fetchAgentList, 
+  updateAgent, 
+  importAgent, 
+  exportAgent, 
+  deleteAgent, 
+  searchAgentInfo 
+} from '@/services/agentConfigService'
+import { generatePromptStream, savePrompt } from '@/services/promptService'
 
 /**
  * Business Logic Configuration Main Component
@@ -680,7 +74,6 @@ export default function BusinessLogicConfig({
 }: BusinessLogicConfigProps) {
   const [enabledToolIds, setEnabledToolIds] = useState<number[]>([]);
   const [isLoadingTools, setIsLoadingTools] = useState(false);
-  const [fileList, setFileList] = useState<any[]>([]);
   const [isImporting, setIsImporting] = useState(false);
   const [isPromptGenerating, setIsPromptGenerating] = useState(false)
   const [isPromptSaving, setIsPromptSaving] = useState(false)
@@ -1484,104 +877,34 @@ export default function BusinessLogicConfig({
             onChange={setBusinessLogic} 
           />
         </div>
-        <div className="w-[280px] flex flex-col self-start">
-          <div className="flex flex-col gap-5 flex-1">
-            <div>
-              <span className="block text-lg font-medium mb-2">{t('businessLogic.config.model')}</span>
-              <Select
-                value={mainAgentModel}
-                onChange={handleModelChange}
-                className="w-full"
-                options={ModelOptions()}
-              />
-            </div>
-            <div>
-              <span className="block text-lg font-medium mb-2">{t('businessLogic.config.maxSteps')}</span>
-              <InputNumber
-                min={1}
-                max={20}
-                value={mainAgentMaxStep}
-                onChange={handleMaxStepChange}
-                className="w-full"
-              />
-            </div>
-            <div className="flex justify-start gap-2 w-full mt-4">
-              <button
-                onClick={() => handleGenerateSystemPrompt(t)}
-                disabled={isPromptGenerating || isPromptSaving}
-                className="px-3.5 py-1.5 rounded-md flex items-center justify-center text-sm bg-blue-500 text-white hover:bg-blue-600 disabled:opacity-50 disabled:cursor-not-allowed"
-                style={{ border: 'none' }}
-              >
-                {isPromptGenerating ? (
-                  <>
-                    <LoadingOutlined spin className="mr-1" />
-                    {t('businessLogic.config.button.generating')}
-                  </>
-                ) : (
-                  <>
-                    <ThunderboltOutlined className="mr-1" />
-                    {t('businessLogic.config.button.generatePrompt')}
-                  </>
-                )}
-              </button>
-              {isCreatingNewAgent && (
-                <>
-                  <button
-                    onClick={handleSaveAsAgent}
-                    disabled={!canSaveAsAgent}
-                    title={getButtonTitle()}
-                    className="px-4 py-1.5 rounded-md flex items-center justify-center text-sm bg-green-500 text-white hover:bg-green-600 disabled:opacity-50 disabled:cursor-not-allowed"
-                    style={{ border: "none" }}
-                  >
-                    {isPromptSaving ? t('businessLogic.config.button.saving') : t('businessLogic.config.button.save')}
-                  </button>
-                  <button
-                    onClick={handleCancelCreating}
-                    disabled={isPromptGenerating || isPromptSaving}
-                    className="px-4 py-1.5 rounded-md flex items-center justify-center text-sm bg-gray-100 text-gray-700 hover:bg-gray-200 disabled:opacity-50 disabled:cursor-not-allowed"
-                    style={{ border: 'none' }}
-                  >
-                    {t('businessLogic.config.button.cancel')}
-                  </button>
-                </>
-              )}
-            </div>
-          </div>
+        <div className="flex flex-col gap-4">
+          <ModelConfigPanel
+            mainAgentModel={mainAgentModel}
+            mainAgentMaxStep={mainAgentMaxStep}
+            onModelChange={handleModelChange}
+            onMaxStepChange={handleMaxStepChange}
+          />
+          <ActionButtons
+            isPromptGenerating={isPromptGenerating}
+            isPromptSaving={isPromptSaving}
+            isCreatingNewAgent={isCreatingNewAgent}
+            canSaveAsAgent={canSaveAsAgent}
+            getButtonTitle={getButtonTitle}
+            onGeneratePrompt={() => handleGenerateSystemPrompt(t)}
+            onSaveAsAgent={handleSaveAsAgent}
+            onCancelCreating={handleCancelCreating}
+          />
         </div>
       </div>
 
       {/* Delete confirmation popup */}
-      <Modal
-        title={t('businessLogic.config.modal.deleteTitle')}
-        open={isDeleteConfirmOpen}
+      <DeleteConfirmModal
+        isOpen={isDeleteConfirmOpen}
+        agentToDelete={agentToDelete}
         onCancel={() => setIsDeleteConfirmOpen(false)}
-        footer={
-          <div className="flex justify-end gap-2">
-            <button 
-              onClick={() => setIsDeleteConfirmOpen(false)}
-              className="px-4 py-1.5 rounded-md flex items-center justify-center text-sm bg-gray-100 text-gray-700 hover:bg-gray-200"
-              style={{ border: "none" }}
-            >
-              {t('businessLogic.config.modal.button.cancel')}
-            </button>
-            <button 
-              onClick={() => handleConfirmDelete(t)}
-              className="px-4 py-1.5 rounded-md flex items-center justify-center text-sm bg-red-500 text-white hover:bg-red-600"
-              style={{ border: "none" }}
-            >
-              {t('businessLogic.config.modal.button.confirm')}
-            </button>
-          </div>
-        }
-        width={400}
-      >
-        <div className="py-4">
-          <Typography.Text>
-            {t('businessLogic.config.modal.deleteContent', { name: agentToDelete?.name })}
-          </Typography.Text>
-        </div>
-      </Modal>
+        onConfirm={() => handleConfirmDelete(t)}
+      />
     </div>
   </TooltipProvider>
   )
-}
+}
